Seperated crypto into seperate classes
